// src/generated/graphql.d.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  BigInt: { input: any; output: any };
  CanvasConfig: { input: any; output: any };
  DateTime: { input: any; output: any };
  DeploymentMeta: { input: any; output: any };
  DisplayConfig: { input: any; output: any };
  EnvironmentConfig: { input: any; output: any };
  EnvironmentVariables: { input: any; output: any };
  JSON: { input: any; output: any };
  RailpackInfo: { input: any; output: any };
  SerializedTemplateConfig: { input: any; output: any };
  ServiceInstanceLimit: { input: any; output: any };
  SubscriptionPlanLimit: { input: any; output: any };
  TemplateConfig: { input: any; output: any };
  TemplateMetadata: { input: any; output: any };
  TemplateServiceConfig: { input: any; output: any };
  TemplateVolume: { input: any; output: any };
  Upload: { input: any; output: any };
};
// ...
export enum ActiveServiceFeatureFlag {
  CopyVolumeToEnvironment = "COPY_VOLUME_TO_ENVIRONMENT",
  Placeholder = "PLACEHOLDER",
}
// ...
export type Environment = Node & {
  __typename?: "Environment";
  createdAt: Scalars["DateTime"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  deploymentTriggers: EnvironmentDeploymentTriggersConnection;
  deployments: EnvironmentDeploymentsConnection;
  id: Scalars["ID"]["output"];
  isEphemeral: Scalars["Boolean"]["output"];
  meta?: Maybe<EnvironmentMeta>;
  name: Scalars["String"]["output"];
  projectId: Scalars["String"]["output"];
  serviceInstances: EnvironmentServiceInstancesConnection;
  sourceEnvironment?: Maybe<Environment>;
  unmergedChangesCount?: Maybe<Scalars["Int"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  variables: EnvironmentVariablesConnection;
  volumeInstances: EnvironmentVolumeInstancesConnection;
};
// ...
export type EnvironmentMeta = {
  __typename?: "EnvironmentMeta";
  baseBranch?: Maybe<Scalars["String"]["output"]>;
  branch?: Maybe<Scalars["String"]["output"]>;
  prCommentId?: Maybe<Scalars["Int"]["output"]>;
  prNumber?: Maybe<Scalars["Int"]["output"]>;
  prRepo?: Maybe<Scalars["String"]["output"]>;
  prTitle?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Project = Node & {
  __typename?: "Project";
  baseEnvironment?: Maybe<Environment>;
  baseEnvironmentId?: Maybe<Scalars["String"]["output"]>;
  botPrEnvironments: Scalars["Boolean"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  deploymentTriggers: ProjectDeploymentTriggersConnection;
  deployments: ProjectDeploymentsConnection;
  description?: Maybe<Scalars["String"]["output"]>;
  environments: ProjectEnvironmentsConnection;
  expiredAt?: Maybe<Scalars["DateTime"]["output"]>;
  groups: ProjectGroupsConnection;
  id: Scalars["ID"]["output"];
  isPublic: Scalars["Boolean"]["output"];
  isTempProject: Scalars["Boolean"]["output"];
  members: Array<ProjectMember>;
  name: Scalars["String"]["output"];
  plugins: ProjectPluginsConnection;
  prDeploys: Scalars["Boolean"]["output"];
  projectPermissions: ProjectProjectPermissionsConnection;
  services: ProjectServicesConnection;
  subscriptionPlanLimit: Scalars["SubscriptionPlanLimit"]["output"];
  subscriptionType: SubscriptionPlanType;
  team?: Maybe<Team>;
  teamId?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  volumes: ProjectVolumesConnection;
  webhooks: ProjectWebhooksConnection;
};
// ...
export type ProjectMember = {
  __typename?: "ProjectMember";
  avatar?: Maybe<Scalars["String"]["output"]>;
  email: Scalars["String"]["output"];
  id: Scalars["String"]["output"];
  name?: Maybe<Scalars["String"]["output"]>;
  role: ProjectRole;
};
// ...
export enum ProjectRole {
  Admin = "ADMIN",
  Member = "MEMBER",
  Viewer = "VIEWER",
}
// ...
export type Service = Node & {
  __typename?: "Service";
  createdAt: Scalars["DateTime"]["output"];
  deletedAt?: Maybe<Scalars["DateTime"]["output"]>;
  deployments: ServiceDeploymentsConnection;
  featureFlags: Array<ActiveServiceFeatureFlag>;
  icon?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  project: Project;
  projectId: Scalars["String"]["output"];
  repoTriggers: ServiceRepoTriggersConnection;
  serviceInstances: ServiceServiceInstancesConnection;
  templateServiceId?: Maybe<Scalars["String"]["output"]>;
  templateThreadSlug?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export enum SubscriptionPlanType {
  Free = "free",
  Hobby = "hobby",
  Pro = "pro",
  Trial = "trial",
}
// ...
export enum UserFlag {
  Beta = "BETA",
}
// ...
export type Volume = Node & {
  __typename?: "Volume";
  createdAt: Scalars["DateTime"]["output"];
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  project: Project;
  projectId: Scalars["String"]["output"];
  volumeInstances: VolumeVolumeInstancesConnection;
};
// ...
export type VolumeInstance = Node & {
  __typename?: "VolumeInstance";
  createdAt: Scalars["DateTime"]["output"];
  currentSizeMB: Scalars["Float"]["output"];
  environment: Environment;
  environmentId: Scalars["String"]["output"];
  externalId?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  mountPath: Scalars["String"]["output"];
  region?: Maybe<Scalars["String"]["output"]>;
  service: Service;
  serviceId?: Maybe<Scalars["String"]["output"]>;
  sizeMB: Scalars["Int"]["output"];
  state?: Maybe<VolumeState>;
  type: VolumeInstanceType;
  volume: Volume;
  volumeId: Scalars["String"]["output"];
};
// ...
export enum VolumeState {
  Deleted = "DELETED",
  Deleting = "DELETING",
  Error = "ERROR",
  Migrating = "MIGRATING",
  MigrationPending = "MIGRATION_PENDING",
  Ready = "READY",
  Restoring = "RESTORING",
  Updating = "UPDATING",
}

export type VolumeUpdateInput = {
  /** The name of the volume */
  name?: InputMaybe<Scalars["String"]["input"]>;
};

export type VolumeVolumeInstancesConnection = {
  __typename?: "VolumeVolumeInstancesConnection";
  edges: Array<VolumeVolumeInstancesConnectionEdge>;
  pageInfo: PageInfo;
};

export type VolumeVolumeInstancesConnectionEdge = {
  __typename?: "VolumeVolumeInstancesConnectionEdge";
  cursor: Scalars["String"]["output"];
  node: VolumeInstance;
};
// ...
export type CreateApiTokenMutation = {
  __typename?: "Mutation";
  apiTokenCreate: string;
};

export type DeleteApiTokenMutationVariables = Exact<{
  id: Scalars["String"]["input"];
}>;

export type DeleteApiTokenMutation = {
  __typename?: "Mutation";
  apiTokenDelete: boolean;
};

export type GetMeQueryVariables = Exact<{ [key: string]: never }>;

export type GetMeQuery = {
  __typename?: "Query";
  me: {
    __typename?: "User";
    name?: string | null;
    isConductor: boolean;
    agreedFairUse: boolean;
    avatar?: string | null;
    banReason?: string | null;
    createdAt: any;
    email: string;
    featureFlags: Array<ActiveFeatureFlag>;
    flags: Array<UserFlag>;
    has2FA: boolean;
    id: string;
    isAdmin: boolean;
    isVerified: boolean;
    lastLogin: any;
    profile?: {
      __typename?: "UserProfile";
      bio?: string | null;
      isPublic: boolean;
      website?: string | null;
    } | null;
    teams: {
      __typename?: "UserTeamsConnection";
      edges: Array<{
        __typename?: "UserTeamsConnectionEdge";
        node: {
          __typename?: "Team";
          id: string;
          name: string;
          preferredRegion?: string | null;
          updatedAt: any;
          createdAt: any;
          avatar?: string | null;
          adoptionLevel: number;
        };
      }>;
    };
  };
};

export type CreateProjectMutationVariables = Exact<{
  teamId: Scalars["String"]["input"];
  name?: InputMaybe<Scalars["String"]["input"]>;
  isPublic?: InputMaybe<Scalars["Boolean"]["input"]>;
  defaultEnvironmentName?: InputMaybe<Scalars["String"]["input"]>;
  description?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type CreateProjectMutation = {
  __typename?: "Mutation";
  projectCreate: {
    __typename?: "Project";
    id: string;
    baseEnvironmentId?: string | null;
    subscriptionType: SubscriptionPlanType;
    subscriptionPlanLimit: any;
    teamId?: string | null;
    name: string;
    isPublic: boolean;
    isTempProject: boolean;
    expiredAt?: any | null;
    createdAt: any;
    baseEnvironment?: {
      __typename?: "Environment";
      id: string;
      createdAt: any;
      name: string;
      updatedAt: any;
      meta?: {
        __typename?: "EnvironmentMeta";
        baseBranch?: string | null;
        branch?: string | null;
        prCommentId?: number | null;
        prNumber?: number | null;
        prRepo?: string | null;
        prTitle?: string | null;
      } | null;
    } | null;
    members: Array<{
      __typename?: "ProjectMember";
      name?: string | null;
      id: string;
      email: string;
      avatar?: string | null;
      role: ProjectRole;
    }>;
  };
};

export type DeleteProjectMutationVariables = Exact<{
  id: Scalars["String"]["input"];
}>;

export type DeleteProjectMutation = {
  __typename?: "Mutation";
  projectDelete: boolean;
};

export type GetProjectQueryVariables = Exact<{
  projectId: Scalars["String"]["input"];
}>;

export type GetProjectQuery = {
  __typename?: "Query";
  project: {
    __typename?: "Project";
    createdAt: any;
    deletedAt?: any | null;
    description?: string | null;
    expiredAt?: any | null;
    id: string;
    isPublic: boolean;
    isTempProject: boolean;
    name: string;
    updatedAt: any;
    services: {
      __typename?: "ProjectServicesConnection";
      edges: Array<{
        __typename?: "ProjectServicesConnectionEdge";
        node: {
          __typename?: "Service";
          createdAt: any;
          deletedAt?: any | null;
          projectId: string;
          id: string;
          name: string;
          icon?: string | null;
          featureFlags: Array<ActiveServiceFeatureFlag>;
          updatedAt: any;
        };
      }>;
    };
  };
};

export type GetProjectsQueryVariables = Exact<{ [key: string]: never }>;

export type GetProjectsQuery = {
  __typename?: "Query";
  projects: {
    __typename?: "QueryProjectsConnection";
    edges: Array<{
      __typename?: "QueryProjectsConnectionEdge";
      node: {
        __typename?: "Project";
        id: string;
        name: string;
        description?: string | null;
        createdAt: any;
        deletedAt?: any | null;
        isPublic: boolean;
        isTempProject: boolean;
        updatedAt: any;
        teamId?: string | null;
      };
    }>;
  };
};

export type CreateServiceMutationVariables = Exact<{
  projectId: Scalars["String"]["input"];
  name?: InputMaybe<Scalars["String"]["input"]>;
  icon?: InputMaybe<Scalars["String"]["input"]>;
  environmentId?: InputMaybe<Scalars["String"]["input"]>;
  branch?: InputMaybe<Scalars["String"]["input"]>;
  image?: InputMaybe<Scalars["String"]["input"]>;
  repo?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type CreateServiceMutation = {
  __typename?: "Mutation";
  serviceCreate: {
    __typename?: "Service";
    createdAt: any;
    featureFlags: Array<ActiveServiceFeatureFlag>;
    icon?: string | null;
    id: string;
    name: string;
    projectId: string;
  };
};
// ...
export type CreateVolumeMutationVariables = Exact<{
  projectId: Scalars["String"]["input"];
  mountPath: Scalars["String"]["input"];
  serviceId?: InputMaybe<Scalars["String"]["input"]>;
  environmentId?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type CreateVolumeMutation = {
  __typename?: "Mutation";
  volumeCreate: {
    __typename?: "Volume";
    volumeInstances: {
      __typename?: "VolumeVolumeInstancesConnection";
      edges: Array<{
        __typename?: "VolumeVolumeInstancesConnectionEdge";
        node: {
          __typename?: "VolumeInstance";
          currentSizeMB: number;
          createdAt: any;
          sizeMB: number;
          serviceId?: string | null;
          region?: string | null;
          id: string;
          mountPath: string;
          externalId?: string | null;
          environmentId: string;
          state?: VolumeState | null;
          type: VolumeInstanceType;
        };
      }>;
    };
  };
};

export type DeleteVolumeMutationVariables = Exact<{
  volumeId: Scalars["String"]["input"];
}>;

export type DeleteVolumeMutation = {
  __typename?: "Mutation";
  volumeDelete: boolean;
};